requirements.txt
django==3.2.5
djangorestframework==3.14.0
traceback2